WebApps/cybersecurity_store/app_sec/requirements.txt
cherrypy
requests
pyotp
cryptography
pycryptodome
